Decreasing the memory allocation by change the hash table
The hash table in origanal construct binary tree
 implementation is used for searching the node index only.
When constructing tree, the function have to allocate extra
memory, which cause the memory waste.

The new implementation give a solution, which is replacing
the "hlist_node" with "Tree_node",and use the linux linked
 list to construct the hash table. In this way, we can use
the "Tree_node" directly while searching the hash table.